Add timestamp

+++ src/main/kotlin/Block.kt
@@ -1,6 +1,6 @@
 data class Block(
         val index: Int,
-        val timestamp: Double,
+        val timestamp: Long,
         val transactions: MutableList<Transaction>?,
         val proof: Int,
         val previousHash: String
